project: adf4377_sdz: migration from mbed to stm32
Project migrated from Mbed(EoL) to STM32 platform.
Its tested and documentation updated.
Kconfig added.
