+++ projects/Kconfig
@@ -62,6 +62,7 @@ source "projects/adin1320/Kconfig"
 source "projects/adf4382/Kconfig"
 source "projects/adf4030/Kconfig"
 source "projects/adf5902_sdz/Kconfig"
+source "projects/adf4377_sdz/Kconfig"
 source "projects/ad7616-st/Kconfig"
 source "projects/ad7768-evb/Kconfig"
 source "projects/ad7768-1fmcz/Kconfig"
