+++ projects/adf4377_sdz/CMakeLists.txt
@@ -1,0 +1,40 @@
+add_executable(adf4377_sdz)
+
+# Common sources
+target_sources(adf4377_sdz PRIVATE
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/common/common_data.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/platform/${PLATFORM}/main.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/platform/${PLATFORM}/parameters.c
+)
+
+# Basic example
+if(CONFIG_ADF4377_BASIC_EXAMPLE)
+    target_sources(adf4377_sdz PRIVATE
+        ${CMAKE_CURRENT_SOURCE_DIR}/src/examples/basic/basic_example.c
+    )
+    target_include_directories(adf4377_sdz PUBLIC
+        ${CMAKE_CURRENT_SOURCE_DIR}/src/examples/basic
+    )
+endif()
+
+# IIO example
+if(CONFIG_ADF4377_IIO_EXAMPLE)
+    target_sources(adf4377_sdz PRIVATE
+        ${CMAKE_CURRENT_SOURCE_DIR}/src/examples/iio/iio_example.c
+    )
+    target_include_directories(adf4377_sdz PUBLIC
+        ${CMAKE_CURRENT_SOURCE_DIR}/src/examples/iio
+    )
+endif()
+
+target_include_directories(adf4377_sdz PUBLIC
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/common
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/platform/${PLATFORM}
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/platform
+)
+
+target_link_libraries(adf4377_sdz no-os)
+
+config_platform_sdk(adf4377_sdz)
+add_flash_target(adf4377_sdz)
+post_build_config(adf4377_sdz)
