+++ projects/adf4377_sdz/Kconfig
@@ -1,0 +1,22 @@
+menu "ADF4377 Configuration"
+
+choice ADF4377_EXAMPLE
+	prompt "Example variant"
+	default ADF4377_BASIC_EXAMPLE
+
+config ADF4377_BASIC_EXAMPLE
+	bool "Basic example"
+	help
+	  Simple example that configures the ADF4377 wideband synthesizer
+	  and generates a fixed RF output frequency.
+
+config ADF4377_IIO_EXAMPLE
+	bool "IIO example"
+	depends on IIO
+	help
+	  IIO framework example. Launches an IIOD server for remote
+	  configuration using IIO Oscilloscope or libiio clients.
+
+endchoice
+
+endmenu
