--- projects/adf4377_sdz/Makefile
@@ -1,14 +1,0 @@
-PLATFORM = mbed
-# PLATFORM = xilinx
-
-# Choose the default example to build
-EXAMPLE ?= iio_example
-
-
-include ../../tools/scripts/generic_variables.mk
-
-include ../../tools/scripts/examples.mk
-
-include src.mk
-
-include ../../tools/scripts/generic.mk
