+++ projects/adf4377_sdz/basic.conf
@@ -1,0 +1,7 @@
+CONFIG_UART=y
+CONFIG_SPI=y
+CONFIG_GPIO=y
+CONFIG_IRQ=y
+CONFIG_FREQUENCY=y
+CONFIG_FREQUENCY_ADF4377=y
+CONFIG_ADF4377_BASIC_EXAMPLE=y
