--- projects/adf4377_sdz/builds.json
@@ -1,10 +1,0 @@
-{
-  "mbed": {
-    "basic_example": {
-      "flags" : "EXAMPLE=basic"
-    },
-    "iio": {
-      "flags": "EXAMPLE=iio_example"
-    }
-  }
-}
