--- projects/adf4377_sdz/builds_ignore.json
@@ -1,8 +1,0 @@
-{
-	"xilinx": {
-	  "demo": {
-		"flags" : "",
-		"hardware" : ["adf4377_sdz_zed"]
-	  }
-	}
-}
