+++ projects/adf4377_sdz/iio.conf
@@ -1,0 +1,8 @@
+CONFIG_UART=y
+CONFIG_SPI=y
+CONFIG_GPIO=y
+CONFIG_IRQ=y
+CONFIG_FREQUENCY=y
+CONFIG_FREQUENCY_IIO_ADF4377=y
+CONFIG_IIO=y
+CONFIG_ADF4377_IIO_EXAMPLE=y
