--- projects/adf4377_sdz/src/examples/iio_example/example.mk
@@ -1,3 +1,0 @@
-IIOD=y
-INCS += $(DRIVERS)/frequency/adf4377/iio_adf4377.h
-SRCS += $(DRIVERS)/frequency/adf4377/iio_adf4377.c
