--- projects/adf4377_sdz/src/platform/mbed/main.c
@@ -1,63 +1,0 @@
-/***************************************************************************//**
- *   @file   main.c
- *   @brief  Main file for Mbed platform of adf4377 project.
- *   @author Antoniu Miclaus (antoniu.miclaus@analog.com)
- *   @author Jude Osemene (jude.osemene@analog.com)
-********************************************************************************
- * Copyright 2025(c) Analog Devices, Inc.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice,
- *    this list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * 3. Neither the name of Analog Devices, Inc. nor the names of its
- *    contributors may be used to endorse or promote products derived from this
- *    software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY ANALOG DEVICES, INC. “AS IS” AND ANY EXPRESS OR
- * IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
- * MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO
- * EVENT SHALL ANALOG DEVICES, INC. BE LIABLE FOR ANY DIRECT, INDIRECT,
- * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA,
- * OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
- * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
- * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
- * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-*******************************************************************************/
-
-#include "parameters.h"
-#include "common_data.h"
-#include "no_os_error.h"
-
-extern int example_main();
-
-/**
- * @brief Main function execution for mbed platform.
- * @return ret - Result of the enabled examples execution.
- */
-int main()
-{
-	int ret = -EINVAL;
-	struct no_os_uart_desc *uart_desc;
-
-	// Initialize UART Interface
-	ret = no_os_uart_init(&uart_desc, &adf4377_uart_ip);
-	if (ret)
-		return ret;
-
-	no_os_uart_stdio(uart_desc);
-
-	// Execute the example main function
-	ret = example_main();
-	if (ret)
-		return ret;
-
-	return ret;
-}
